@robodoo 2025-08-22
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/223875.png)](https://mergebot.odoo.com/odoo/odoo/pull/223875)